+++ src/entities/Camera.cpp
@@ -328,6 +328,90 @@ void Camera::notify_size_changed() {
 }
 
 /**
+ * \copydoc Entity::is_block_obstacle
+ */
+bool Camera::is_block_obstacle(Block&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_teletransporter_obstacle
+ */
+bool Camera::is_teletransporter_obstacle(Teletransporter&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_stream_obstacle
+ */
+bool Camera::is_stream_obstacle(Stream&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_stairs_obstacle
+ */
+bool Camera::is_stairs_obstacle(Stairs&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_switch_obstacle
+ */
+bool Camera::is_switch_obstacle(Switch&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_raised_block_obstacle
+ */
+bool Camera::is_raised_block_obstacle(CrystalBlock&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_crystal_obstacle
+ */
+bool Camera::is_crystal_obstacle(Crystal&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_npc_obstacle
+ */
+bool Camera::is_npc_obstacle(Npc&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_door_obstacle
+ */
+bool Camera::is_door_obstacle(Door&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_jumper_obstacle
+ */
+bool Camera::is_jumper_obstacle(Jumper&, const Rectangle&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_destructible_obstacle
+ */
+bool Camera::is_destructible_obstacle(Destructible&) {
+  return false;
+}
+
+/**
+ * \copydoc Entity::is_chest_obstacle
+ */
+bool Camera::is_chest_obstacle(Chest&) {
+  return false;
+}
+
+/**
  * \copydoc Entity::is_separator_obstacle
  */
 bool Camera::is_separator_obstacle(Separator& separator, const Rectangle& candidate_position) {