+++ src/entities/Chest.cpp
@@ -357,8 +357,8 @@ void Chest::set_cannot_open_dialog_id(const std::string& cannot_open_dialog_id) 
  * \param other Another entity.
  * \return \c true if this entity is an obstacle for the other one.
  */
-bool Chest::is_obstacle_for(Entity& /* other */) {
-  return true;
+bool Chest::is_obstacle_for(Entity& other) {
+  return other.is_chest_obstacle(*this);
 }
 
 /**