+++ tests/cmake/AddTestMaps.cmake
@@ -194,6 +194,7 @@ list(APPEND LUA_TEST_MAPS
   "bugs/1609_crash_save_lua_keyword"
   "bugs/1611_entity_overlaps_when_removed"
   "bugs/1617_crash_disabled_chest"
+  "bugs/1619_camera_obstacles"
 )
 
 # Test that should succeed when there are non-fatal errors