+++ changelog.md
@@ -18,6 +18,7 @@
 * Fix `hero.set_invincible` not suspended with the game (#1554).
 * Fix custom entity still interacting when moved or disabled (#1056, #1407).
 * Fix thrown objects destruction animation position (#1508).
+* Fix camera stuck on entities other than separators (#1619).
 * Fix crash when loading a save file with Lua keywords (#1609).
 * Fix crash when a dialog starts during another dialog (#988, #1582).